Refactor(#17): 페이지 내부 타입을 분리 리팩토링
