+++ src/app/(auth)/sign-in/phone/type.ts
@@ -1,0 +1,5 @@
+export interface PhoneSignInForm {
+  phone: string;
+  password: string;
+  remember: boolean;
+}
